ahci: update cmd_tbl and rx_fis to 64-bit
existing u32 datatypes may fail to use correct value
when mem pointer crosses 32-bit boundary.

+++ include/ahci.h
@@ -137,8 +137,8 @@ struct ahci_ioports {
 	void __iomem	*port_mmio;
 	struct ahci_cmd_hdr	*cmd_slot;
 	struct ahci_sg		*cmd_tbl_sg;
-	ulong	cmd_tbl;
-	u32	rx_fis;
+	u64	cmd_tbl;
+	u64	rx_fis;
 };
 
 /**
